+++ packages/core/src/index.ts
@@ -31,11 +31,14 @@ export { Observable, Subscription } from './tools/observable'
 export {
   startSessionManagement,
   Session,
+  // Exposed for tests
+  stopSessionManagement,
+} from './domain/session/sessionManagement'
+export {
   SESSION_TIME_OUT_DELAY,
   // Exposed for tests
   SESSION_COOKIE_NAME,
-  stopSessionManagement,
-} from './domain/sessionManagement'
+} from './domain/session/sessionStore'
 export { HttpRequest, Batch, canUseEventBridge, getEventBridge } from './transport'
 export * from './tools/display'
 export * from './tools/urlPolyfill'
@@ -46,7 +49,7 @@ export * from './tools/browserDetection'
 export { instrumentMethod, instrumentMethodAndCallOriginal } from './tools/instrumentMethod'
 export { ErrorSource, ErrorHandling, formatUnknownError, createHandlingStack, RawError } from './tools/error'
 export { Context, ContextArray, ContextValue } from './tools/context'
-export { areCookiesAuthorized, getCookie, setCookie, COOKIE_ACCESS_DELAY } from './browser/cookie'
+export { areCookiesAuthorized, getCookie, setCookie, deleteCookie, COOKIE_ACCESS_DELAY } from './browser/cookie'
 export { initXhrObservable, XhrCompleteContext, XhrStartContext } from './browser/xhrObservable'
 export { initFetchObservable, FetchCompleteContext, FetchStartContext, FetchContext } from './browser/fetchObservable'
 export { EndpointBuilder } from './domain/configuration/endpointBuilder'